fix: close PR 5756 audit regressions
